Added more sprites, fixed box2d world offset

+++ core/src/com/portal2d/game/controller/Box2DConstants.java
@@ -8,7 +8,7 @@ import com.badlogic.gdx.math.Vector2;
 public interface Box2DConstants {
 
     // scale for box2d world, pixels per meter
-    int PPM = 100;
+    float PPM = 100;
 
     Vector2 DEFAULT_GRAVITY = new Vector2(0, -9.8f);
